# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talent Bridge API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/certificates": {
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Create a new certificate",
                "operationId": "a00c8fcda1fd26ec6c1e4aff4cbe3395",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "certificate_name",
                                    "expiration_date",
                                    "category",
                                    "status",
                                    "file_path",
                                    "description"
                                ],
                                "properties": {
                                    "certificate_name": {
                                        "type": "string",
                                        "example": "AWS Certified Developer"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Cloud Computing"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "example": "certificates/aws-dev-cert.pdf"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Certification for AWS developer associate level"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Certificate created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Certificate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "invalid add certificate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/certificates/{id}": {
            "put": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Update a certificate",
                "operationId": "5eaff6015aa0dcc0fa43913eac436ac8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Certificate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "certificate_name",
                                    "expiration_date",
                                    "category",
                                    "status",
                                    "file_path",
                                    "description"
                                ],
                                "properties": {
                                    "certificate_name": {
                                        "type": "string",
                                        "example": "Updated AWS Certified Developer"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Cloud Computing"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "example": "certificates/updated-aws-dev-cert.pdf"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated certification for AWS developer associate level"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Certificate updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ststus": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Certificate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "invalid add certificate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "massage": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Delete a certificate",
                "operationId": "b3ee9a288a3aa567a708f702492896fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Certificate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "delete succesfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Certificate not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "massage": {
                                            "type": "string",
                                            "example": "certificate is nothing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Certificate": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "certificate_name": {
                        "type": "string",
                        "example": "AWS Certified Developer"
                    },
                    "expiration_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "category": {
                        "type": "string",
                        "example": "Cloud Computing"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "file_path": {
                        "type": "string",
                        "example": "certificates/aws-dev-cert.pdf"
                    },
                    "description": {
                        "type": "string",
                        "example": "Certification for AWS developer associate level"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}